# public/swagger.json
{
  "swagger": "2.0",
  "info": {
    "contact": {
      "email": "marius.klocke@gmail.com",
      "name": "Marius Klocke"
    },
    "description": "RESTful API for managing Wilde Liga Bremen",
    "title": "Wilde Liga Bremen API",
    "version": "0.1"
  },
  "basePath": "/api",
  "paths": {
...
    "/tournament/{id}/round/{round}": {
      "put": {
        "operationId": "setRound",
        "tags": [
          "tournament"
        ],
        "parameters": [
          {
            "description": "ID of tournament",
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          },
          {
            "description": "Number identifying the tournament round",
            "in": "path",
            "name": "round",
            "required": true,
            "type": "integer"
          },
          {
            "in": "body",
            "name": "setRoundBody",
            "schema": {
              "type": "object",
              "properties": {
                "team_pairs": {
                  "type": "array",
                  "items": {
                    "properties": {
                      "home_team_id": {
                        "type": "string"
                      },
                      "guest_team_id": {
                        "type": "string"
                      }
                    },
                    "type": "object"
                  },
                  "minItems": 1,
                  "maxItems": 64
                },
                "planned_for": {
                  "type": "string",
                  "format": "date"
                }
              }
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Operation successful"
          },
          "404": {
            "description": "Tournament or team not found"
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "basic-auth": {
      "type": "basic"
    },
    "JWT": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
...
}